requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
certifi>=2024.0.0
cryptography>=42.0.0
